+++ test_apps/esp-sr/main/test_wakenet.c
@@ -94,14 +94,16 @@ TEST_CASE("wakenet detect API & cpu loading", "[wn]")
     gettimeofday(&tv_start, NULL);
     unsigned char* data = NULL;
     size_t data_size = 0;
+    char *wake_words = NULL;
+    wake_words = esp_srmodel_get_wake_words(models, model_name);
     if (strstr(model_name, "hiesp") != NULL) {
         data = (unsigned char*)hiesp;
         data_size = sizeof(hiesp);
-        printf("wake word: Hi, ESP, size:%d\n", data_size);
+        printf("wake word: %s, size:%d\n", wake_words,  data_size);
     } else if(strstr(model_name, "hilexin") != NULL) {
         data = (unsigned char*)hilexin;
         data_size = sizeof(hilexin);
-        printf("wake word: hi,lexin, size:%d\n", data_size);
+        printf("wake word: %s, size:%d\n", wake_words,  data_size);
     }
 
     while (1) {